Added Server Time Feature
Server time extraction is now available via the -d flag on the map_reader.py script.
It will access the server via telnet to retrieve the time and store in a DateTime.csv
file readable by the leaflet javascript. The time will be added to an overlay box at
top left of the map.